refactor: Update import paths for navbar components
