+++ src/components/admin-panel/navbar/mode-toggle.tsx
@@ -1,4 +1,3 @@
-"use client";
 import { useTheme } from "next-themes";
 import { MoonIcon, SunIcon } from "@radix-ui/react-icons";
