+++ src/components/admin-panel/navbar/sheet-menu.tsx
@@ -2,7 +2,7 @@ import Link from "next/link";
 import { MenuIcon, PanelsTopLeft } from "lucide-react";
 
 import { Button } from "@/components/ui/button";
-import { Menu } from "@/components/admin-panel/menu";
+import { Menu } from "@/components/admin-panel/sidebar-menu/menu";
 import {
   Sheet,
   SheetHeader,
