+++ src/components/admin-panel/navbar/user-nav.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import Link from "next/link";
 import { LayoutGrid, LogOut, User } from "lucide-react";
 
@@ -9,7 +7,7 @@ import {
   Tooltip,
   TooltipContent,
   TooltipTrigger,
-  TooltipProvider
+  TooltipProvider,
 } from "@/components/ui/tooltip";
 import {
   DropdownMenu,
@@ -18,7 +16,7 @@ import {
   DropdownMenuItem,
   DropdownMenuLabel,
   DropdownMenuSeparator,
-  DropdownMenuTrigger
+  DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 
 export function UserNav() {
